server 모듈화
